requirements.txt
langchain>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langchain-community==1.0.0a1
langchain-openai>=0.2.0
pydantic>=2.6
requests>=2.31
tenacity>=8.2
numpy>=1.24
pandas>=2.0
python-dateutil>=2.8
tqdm>=4.66
pytest>=7.4
pytest-mock>=3.10
yfinance>=0.2
openai>=1.52
python-dotenv>=1.0
